fix: linter error
